commit result

+++ src/main/java/com/example/ProjectBoot1/dao/UserDaoImpl.java
@@ -2,7 +2,6 @@ package com.example.ProjectBoot1.dao;
 
 import com.example.ProjectBoot1.model.User;
 import org.springframework.stereotype.Repository;
-
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import java.util.List;
@@ -11,44 +10,44 @@ import java.util.List;
 @Repository
 public class UserDaoImpl implements UserDao {
 
-
     @PersistenceContext
     private EntityManager entityManager;
 
     @Override
-    public List<User> getUsers() {
-        return entityManager.createQuery("SELECT u FROM User u", User.class).getResultList();
+    public List<User> getAllUsers() {
+        return entityManager.createQuery("select u from User u", User.class).getResultList();
     }
 
     @Override
-
-    public User getUserById(long id) {
+    public User getUserById(Long id) {
         return entityManager.find(User.class, id);
-
     }
 
     @Override
     public void addUser(User user) {
         entityManager.persist(user);
-        entityManager.flush();
+    }
+
+    @Override
+    public void deleteUser(Long id) {
+        try {
+            User user = entityManager.find(User.class, id);
+            if (user != null) {
+                entityManager.remove(user);
+            }
+        } catch (NullPointerException e) {
+            System.out.println("User с указанным вами id не существует!");
+        }
     }
 
     @Override
     public void updateUser(User user) {
         entityManager.merge(user);
-
-    }
-
-    @Override
-    public void deleteUser(long id) {
-        User user = entityManager.find(User.class, id);
-        entityManager.remove(user);
-        entityManager.flush();
     }
 
     @Override
     public User getUserByName(String name) {
-        return entityManager.createQuery("SELECT u FROM User u where u.name = :name", User.class)
+        return entityManager.createQuery("select u from User u where u.name = :name", User.class)
                 .setParameter("name", name)
                 .getSingleResult();
     }
